requirements.txt
# Core ML Libraries
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Vector Search
faiss-cpu>=1.7.4

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
